+++ Aurora/External/libRocket/Source/Core/PropertySpecification.cpp
@@ -14,7 +14,7 @@
  *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -313,7 +313,7 @@ bool PropertySpecification::ParsePropertyValues(StringList& values_list, const S
 	ParseState state = VALUE;
 	int open_parentheses = 0;
 
-	size_t character_index = 0;
+	uint32_t character_index = 0;
 	char previous_character = 0;
 	while (character_index < values.Length())
 	{
